Fix SDK lib exports
