+++ lib/src/SDK/SDK.c
@@ -1,5 +1,7 @@
 // OpenCL SDK includes
 #include <CL/SDK/SDK.h>
 
+#include "CLI.c"
 #include "Context.c"
 #include "Image.c"
+#include "Random.c"
