+++ lib/src/SDK/SDK.cpp
@@ -1,6 +1,9 @@
 // OpenCL SDK includes
 #include <CL/SDK/SDK.hpp>
 
+#include "CLI.cpp"
 #include "Image.cpp"
+#ifdef OPENCL_SDK_BUILD_OPENGL_SAMPLES
 #include "InteropContext.cpp"
 #include "InteropWindow.cpp"
+#endif
